+++ src/Ydb.Sdk/src/Ado/PoolManager.cs
@@ -20,10 +20,10 @@ internal static class PoolManager
             return await sessionPool.OpenSession(cancellationToken);
         }
 
+        await SemaphoreSlim.WaitAsync(cancellationToken);
+
         try
         {
-            await SemaphoreSlim.WaitAsync(cancellationToken);
-
             if (Pools.TryGetValue(settings.ConnectionString, out var pool))
             {
                 return await pool.OpenSession(cancellationToken);
@@ -62,12 +62,11 @@ internal static class PoolManager
 
     internal static async Task ClearPool(string connectionString)
     {
-        if (Pools.Remove(connectionString, out var sessionPool))
+        if (Pools.TryRemove(connectionString, out var sessionPool))
         {
+            await SemaphoreSlim.WaitAsync();
             try
             {
-                await SemaphoreSlim.WaitAsync();
-
                 await sessionPool.DisposeAsync();
             }
             finally