+++ src/EFCore.Ydb/CHANGELOG.md
@@ -1,3 +1,4 @@
+- Fixed Decimal precision/scale mapping in EF provider.
 - Supported Guid (Uuid YDB type).
 - PrivateAssets="none" is set to flow the EF Core analyzer to users referencing this package [issue](https://github.com/aspnet/EntityFrameworkCore/pull/11350).